covidapp/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.2
django-bootstrap4==3.0.1
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
